Add files via upload
In this module, we have to --
i) implement the isLand(Feature earthquake) method in EarthquakeCityMap. This method should return true if the location of the input earthquake is on land. It should also set the "country" property on the LandMarker to the country where the earthquake occurred.
ii) Implement the printQuakes() method in EarthquakeCityMap.
iii) Complete the method definition for the draw() method in the CityMarker class. 
iv) Complete the method definitions for the drawEarthquake() method in LandQuakeMarker and OceanQuakeMarker so that the size of the marker varies by its magnitude (larger magnitude=larger marker).
v) Complete the method definition for colorDetermine() method in EarthquakeMarker